Updated comments on previous pr  - changed AvatarSkillData.java to read costElemVal as float, so that can use directly by PropUpdate packets  - updated Avatar.java to not use hardcoded depot values  - added function in EntityAvatar.java to add energy to only the active entity avatar  - added function in EnergyManager.java to only refill energy for the active entity avatar  - updated changeAvatarElement function in Player.java to be bit more robust, also sending packets that mimics the official server  - updated ExecAddCurAvatarEnergy.java to use return value directly from refill energy function  - changed HandlerAvatarChangeElementTypeReq.java to use changeAvatarElement function in Player.java  - changed PacketAvatarFightPropUpdate.java to accept list of fight prop as well  -

+++ src/main/java/emu/grasscutter/data/excels/AvatarSkillData.java
@@ -10,7 +10,7 @@ import lombok.Getter;
 public class AvatarSkillData extends GameResource {
     private int id;
     @Getter private float cdTime;
-    @Getter private int costElemVal;
+    @Getter private float costElemVal; // fight prop update uses float
     @Getter private int maxChargeNum;
     @Getter private int triggerID;
     @Getter private boolean isAttackCameraLock;
